If test failed, raise
