+++ lib/paraduct.rb
@@ -6,6 +6,7 @@ require "paraduct/configuration"
 require "paraduct/variable_converter"
 require "paraduct/runner"
 require "paraduct/parallel_runner"
+require "paraduct/errors"
 require "pathname"
 
 module Paraduct
