+++ lib/paraduct/runner.rb
@@ -1,17 +1,6 @@
 module Paraduct
   require "open3"
 
-  class ProcessError < StandardError
-    attr_reader :status
-
-    # @param message [String] stdout and stderr
-    # @param status  [Process::Status]
-    def initialize(message, status)
-      super(message)
-      @status = status
-    end
-  end
-
   class Runner
     # run script with params
     # @param script [String, Array<String>] script file, script(s)
