+++ spec/.rubocop.yml
@@ -1,0 +1,5 @@
+inherit_from:
+  - ../.hound.yml
+
+Style/Blocks:
+  Enabled: false
